requirements.txt
firebase_admin==7.0.0
pandas==2.3.1
plotly==6.2.0
Pyrebase4==4.8.0
python-dotenv==1.1.1
SQLAlchemy==2.0.41
streamlit==1.46.1